app2/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
attrs==24.2.0
    # via
    #   jsonschema
    #   referencing
bcrypt==4.2.1
    # via streamlit-authenticator
blinker==1.9.0
    # via streamlit
cachetools==5.5.0
    # via streamlit
captcha==0.6.0
    # via streamlit-authenticator
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via streamlit
colorama==0.4.6
    # via click
extra-streamlit-components==0.1.71
    # via streamlit-authenticator
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
idna==3.10
    # via requests
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2024.10.1
    # via jsonschema
lxml==5.3.0
    # via python-docx
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
narwhals==1.15.2
    # via altair
numpy==2.1.3
    # via
    #   pandas
    #   pydeck
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   streamlit
pandas==2.2.3
    # via streamlit
pillow==11.0.0
    # via
    #   detached-secxxii (pyproject.toml)
    #   captcha
    #   streamlit
protobuf==5.29.1
    # via streamlit
pyarrow==18.1.0
    # via streamlit
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
pyjwt==2.10.1
    # via streamlit-authenticator
python-dateutil==2.9.0.post0
    # via pandas
python-docx==1.1.2
    # via detached-secxxii (pyproject.toml)
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   detached-secxxii (pyproject.toml)
    #   streamlit-authenticator
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via streamlit
rich==13.9.4
    # via streamlit
rpds-py==0.22.3
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.40.2
    # via
    #   detached-secxxii (pyproject.toml)
    #   extra-streamlit-components
    #   streamlit-authenticator
    #   streamlit-calendar
    #   streamlit-option-menu
streamlit-authenticator==0.4.1
    # via detached-secxxii (pyproject.toml)
streamlit-calendar==1.2.0
    # via detached-secxxii (pyproject.toml)
streamlit-option-menu==0.4.0
    # via detached-secxxii (pyproject.toml)
tenacity==9.0.0
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   python-docx
    #   streamlit
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
watchdog==6.0.0
    # via streamlit
